Move masonry grid to AdobeUi module (prepare for moving to magento2)
