+++ AdobeStockImageAdminUi/view/adminhtml/web/js/components/images-grid-sizes.js
