+++ AdobeUi/composer.json
@@ -1,11 +1,9 @@
 {
-    "name": "magento/module-ui",
+    "name": "magento/module-adobe-ui",
     "description": "N/A",
     "require": {
         "php": "~7.1.3||~7.2.0",
-        "magento/framework": "*",
-        "magento/module-adobe-stock-asset-api": "*",
-        "magento/module-adobe-stock-image-api": "*"
+        "magento/framework": "*"
     },
     "type": "magento2-module",
     "license": [
